updated job_start options to allow job to finish
The validator.vim plugin causes a segfault if a file is being validated
at the time :wq is run. Tracing the issue appears to the show the
SIGSEGV occurs when new_threadstate is called from the Python libraries.
My rough suspicion is that the SIGTERM sent to the linter (itself
running Python) may be cause. Preventing the SIGTERM is what this
commit does, which resolves maralla/validator.vim#53.